+++ fs/debugfs/cxl/mbox/raw_allow_all.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+set -e
+
+sudo cat /sys/kernel/debug/cxl/mbox/raw_allow_all