+++ fs/debugfs/cxl/mem0/inject_poison.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+# /sys/kernel/debug/cxl/mem0/inject_poison
+# /sys/kernel/debug/cxl/mem0/clear_poison
+#
+# $ sudo cxl {inject-media-poison,inject-protocol-error}
+# may use this file.
+set -e